Fix following splint warnigs.
src/elab.c: (in function find_arch)
src/elab.c:88:28: Variable new_line initialized to type uint16_t,
expects int:
                     tree_loc(t)->first_line
  To allow arbitrary integral types to match any integral type, use
  +matchanyintegral.
src/elab.c:89:28: Variable old_line initialized to type uint16_t,
expects int:
                     tree_loc(*(params->arch))->first_line

+++ src/elab.c
@@ -79,8 +79,8 @@ static void find_arch(tree_t t, void *context)
             // Note this assumes both architectures are from the same
             // file but this shouldn't be a problem with high-resolution
             // timestamps
-            int new_line = tree_loc(t)->first_line;
-            int old_line = tree_loc(*(params->arch))->first_line;
+            uint16_t new_line = tree_loc(t)->first_line;
+            uint16_t old_line = tree_loc(*(params->arch))->first_line;
 
             if (new_line > old_line)
                *(params->arch) = t;
